fix: Add record_global.go
