+++ casdoorsdk/record_global.go
@@ -1,4 +1,4 @@
-// Copyright 2021 The Casdoor Authors. All Rights Reserved.
+// Copyright 2023 The Casdoor Authors. All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,6 +14,18 @@
 
 package casdoorsdk
 
-func SendSms(content string, receivers ...string) error {
-	return globalClient.SendSms(content, receivers...)
+func GetRecords() ([]*Record, error) {
+	return globalClient.GetRecords()
+}
+
+func GetPaginationRecords(p int, pageSize int, queryMap map[string]string) ([]*Record, int, error) {
+	return globalClient.GetPaginationRecords(p, pageSize, queryMap)
+}
+
+func GetRecord(name string) (*Record, error) {
+	return globalClient.GetRecord(name)
+}
+
+func AddRecord(record *Record) (bool, error) {
+	return globalClient.AddRecord(record)
 }
